A major overhaul of GeneralFeature parser tests.
